refactor: keep legacy spacing-index conversion only for Flyway replay
Move hoursFromLegacyIndex onto StabilityIndexToHoursBackfill so V300000260
can still replay. Delete live SpacedRepetitionAlgorithm. FSRS-6 short-term
under one day is now in product code and Proposed ADR 0003.
