+++ .planning/research/FSRS-COMPATIBILITY-GAP.md
@@ -14,7 +14,7 @@ Product policy lives in ADR 0003 Decision. This tracker is a pointer plus the de
 
 Doughnut persists **Stability** in whole hours and **Difficulty** (nullable; shown on the Memory Tracker). Retrievability is computed (FSRS-6 power curve), not stored. Frozen default FSRS-6 weights live in `Fsrs`. Requested retention is locked global `r = 0.9` (`Fsrs.REQUESTED_RETENTION`). There is **no lapse count**. Maximum interval is locked global **36500 days** / **876000 hours** (`Fsrs.MAXIMUM_INTERVAL_HOURS`); clamp after next-S, due from that S. There is **no interval fuzz** (due is `lastRecalledAt + I(0.9, S)`). Thinking time is recorded on answers for display and stats; it is not a memory-state input. There is no FSRS card-state machine (Learning / Review / Relearning); Doughnut **New** is ungraded. Assimilate does not write `lastRecalledAt`; New due is `assimilatedAt`. `lastRecalledAt` is the last mapped grade. First-rating on New uses published FSRS-6 `S0(G)` / `D0(G)` for all four G (Tutor **2** on New is Hard). RecallLog `elapsed_hours` is required (ADR 0003 RecallLog). Graded success (Hard/Good/Easy) uses published FSRS-6 short-term next Stability when elapsed whole hours are **< 24**, and long-term Stability increase when elapsed is **≥ 24**.
 
-Live scheduling does not walk a spacing-index ladder. `DEFAULT_SPACES` / `hoursFromLegacyIndex` remain only so committed `V300000260` can replay on fresh DBs.
+Live scheduling does not walk a spacing-index ladder. Legacy index conversion (`hoursFromLegacyIndex`) lives only on `StabilityIndexToHoursBackfill` so committed `V300000260` can replay on fresh DBs.
 
 ## Deferred (see ADR 0003 Decision)
 
@@ -28,6 +28,6 @@ Humans still own accept / reject / supersede of ADR 0003 (`docs/adrs/README.md`)
 - [ADR 0001](../../docs/adrs/0001-ubiquitous-language.md) — **recall** vs FSRS **review**; **New** = ungraded
 - [ADR 0005](../../docs/adrs/0005-commissioned-learning-session-protocol.md) — Tutor 1–4 meaning (`score = G`)
 - Seed: [SEED-004](../seeds/SEED-004-close-spaced-repetition-scheduling-policy-gap.md)
-- Code: `Fsrs`, `MemoryTracker`, `SpacedRepetitionAlgorithm`
+- Code: `Fsrs`, `MemoryTracker`; Flyway replay only: `StabilityIndexToHoursBackfill` / `V300000260`
 - [FSRS-6 algorithm](https://github.com/open-spaced-repetition/awesome-fsrs/wiki/The-Algorithm)
 - [ts-fsrs](https://open-spaced-repetition.github.io/ts-fsrs/)
